Make archive methods async
This is probably an extremely bad idea, like, with regard to
performance, splitting every functions in as many piece of codes as
there are `read` and `write` operations is probably going to confuse the
optimizer so much it's not funny.

I also removed the GNU extension related code because I didn't need it.